Added Grand wizard unique attack

+++ SpellingTestUnity/Assets/Scripts/FireWizard.cs
@@ -22,10 +22,11 @@ public class FireWizard : MonoBehaviour
     private bool isAttacking = false;
     private float attackTimer;
     private float attackCd = 1f;
-    private float basicAttacksCount = 0;
+    private int basicAttacksCount = 0;
+
     private float uniqueAttackTimer;
     private float uniqueAttackCd = 0.3f;
-    private float uniqueAttackCount = 0;
+    private int uniqueAttackCount = 0;
 
     private List<KeyValuePair<GameObject, Vector2>> attackParticlesList;
